+++ lib/archethic/contracts/interpreter/version0/utils_interpreter.ex
@@ -1,9 +1,9 @@
-defmodule Archethic.Contracts.Interpreter.Utils do
+defmodule Archethic.Contracts.Interpreter.Version0.UtilsInterpreter do
   @moduledoc false
 
   alias Archethic.Crypto
-  alias Archethic.Contracts.Interpreter.Library
-  alias Archethic.Contracts.Interpreter.TransactionStatements
+  alias Archethic.Contracts.Interpreter.Version0.Library
+  alias Archethic.Contracts.Interpreter.Version0.TransactionStatements
 
   @library_functions_names Library.__info__(:functions)
                            |> Enum.map(&Atom.to_string(elem(&1, 0)))
@@ -536,56 +536,6 @@ defmodule Archethic.Contracts.Interpreter.Utils do
   end
 
   defp do_postwalk_execution(node, acc), do: {node, acc}
-
-  @doc """
-  Format an error message from the failing ast node
-
-  It returns message with metadata if possible to indicate the line of the error
-  """
-  @spec format_error_reason(any(), String.t()) :: String.t()
-  def format_error_reason({:atom, _key}, reason) do
-    do_format_error_reason(reason, "", [])
-  end
-
-  def format_error_reason({{:atom, key}, metadata, _}, reason) do
-    do_format_error_reason(reason, key, metadata)
-  end
-
-  def format_error_reason({_, metadata, [{:__aliases__, _, [atom: module]} | _]}, reason) do
-    do_format_error_reason(reason, module, metadata)
-  end
-
-  def format_error_reason(ast_node = {_, metadata, _}, reason) do
-    do_format_error_reason(reason, Macro.to_string(ast_node), metadata)
-  end
-
-  def format_error_reason({{:atom, _}, {_, metadata, _}}, reason) do
-    do_format_error_reason(reason, "", metadata)
-  end
-
-  def format_error_reason({{:atom, key}, _}, reason) do
-    do_format_error_reason(reason, key, [])
-  end
-
-  defp do_format_error_reason(message, cause, metadata) do
-    message = prepare_message(message)
-
-    [prepare_message(message), cause, metadata_to_string(metadata)]
-    |> Enum.reject(&(&1 == ""))
-    |> Enum.join(" - ")
-  end
-
-  defp prepare_message(message) when is_atom(message) do
-    message |> Atom.to_string() |> String.replace("_", " ")
-  end
-
-  defp prepare_message(message) when is_binary(message) do
-    String.trim_trailing(message, ":")
-  end
-
-  defp metadata_to_string(line: line, column: column), do: "L#{line}:C#{column}"
-  defp metadata_to_string(line: line), do: "L#{line}"
-  defp metadata_to_string(_), do: ""
 
   @doc """
   Gets an valid address as true binary from string term or from true binary <<>>.
