+++ test/archethic/contracts/interpreter/version0/library_test.exs
@@ -1,7 +1,7 @@
-defmodule Archethic.Contracts.Interpreter.LibraryTest do
+defmodule Archethic.Contracts.Interpreter.Version0.LibraryTest do
   use ArchethicCase
 
-  alias Archethic.Contracts.Interpreter.Library
+  alias Archethic.Contracts.Interpreter.Version0.Library
 
   alias Archethic.P2P.Message.GetFirstTransactionAddress
   alias Archethic.P2P.Message.FirstTransactionAddress
