+++ test/archethic/contracts/interpreter/version0/transaction_statements_test.exs
@@ -1,8 +1,8 @@
-defmodule Archethic.Contracts.Interpreter.TransactionStatementsTest do
+defmodule Archethic.Contracts.Interpreter.Version0.TransactionStatementsTest do
   @moduledoc false
   use ExUnit.Case
 
-  alias Archethic.Contracts.Interpreter.TransactionStatements
+  alias Archethic.Contracts.Interpreter.Version0.TransactionStatements
 
   alias Archethic.TransactionChain.Transaction
   alias Archethic.TransactionChain.TransactionData
