+++ test/archethic/contracts/interpreter/version1_test.exs
@@ -1,0 +1,8 @@
+defmodule Archethic.Contracts.Interpreter.Version1Test do
+  @moduledoc false
+  use ArchethicCase
+
+  alias Archethic.Contracts.Interpreter.Version1
+
+  doctest Version1
+end
